+++ src/transforms/serializer.js
@@ -6,35 +6,24 @@ const Parser = require("protodef").Parser;
 
 const minecraft = require("../datatypes/minecraft");
 const states = require("../states");
+const merge=require("lodash.merge");
+const get=require("lodash.get");
 
-function createProtocol(types,packets)
+function recursiveAddTypes(protocol,protocolData,path)
+{
+  if(protocolData===undefined)
+    return;
+  if(protocolData.types)
+    protocol.addTypes(protocolData.types);
+  recursiveAddTypes(protocol,get(protocolData,path.shift()),path);
+}
+
+function createProtocol(state,direction,version,customPackets)
 {
   const proto = new ProtoDef();
-  proto.addType("string",["pstring",{
-    countType:"varint"
-  }]);
   proto.addTypes(minecraft);
-  proto.addTypes(types);
-
-  Object.keys(packets).forEach(function(name) {
-    proto.addType("packet_"+name,["container",packets[name].fields]);
-  });
-
-  proto.addType("packet",["container", [
-    { "name": "name", "type":["mapper",{"type": "varint" ,
-      "mappings":Object.keys(packets).reduce(function(acc,name){
-        acc[parseInt(packets[name].id)]=name;
-        return acc;
-      },{})
-    }]},
-    { "name": "params", "type": ["switch", {
-      "compareTo": "name",
-      "fields": Object.keys(packets).reduce(function(acc,name){
-        acc[name]="packet_"+name;
-        return acc;
-      },{})
-    }]}
-  ]]);
+  const mcData=require("minecraft-data")(version);
+  recursiveAddTypes(proto,merge(mcData.protocol,get(customPackets,[mcData.version.majorVersion])),[state,direction]);
   return proto;
 }
 
@@ -44,11 +33,12 @@ function createSerializer(opts)
   const state = opts.state !== undefined ? opts.state : states.HANDSHAKING;
   const isServer = opts.isServer !== undefined ? opts.isServer : false;
   const version = opts.version;
+  const customPackets = opts.customPackets;
 
   const mcData=require("minecraft-data")(version);
   const direction = !isServer ? 'toServer' : 'toClient';
   const packets = mcData.protocol.states[state][direction];
-  const proto=createProtocol(mcData.protocol.types,packets);
+  const proto=createProtocol(mcData.protocol.types,packets,customPackets);
   return new Serializer(proto,"packet");
 }
 
@@ -59,13 +49,13 @@ function createDeserializer(opts)
   const isServer = opts.isServer !== undefined ? opts.isServer : false;
   const packetsToParse = opts.packetsToParse !== undefined ? opts.packetsToParse : {"packet": true};
   const version = opts.version;
-
+  const customPackets = opts.customPackets;
 
   const mcData=require("minecraft-data")(version);
   const direction = isServer ? "toServer" : "toClient";
   const packets = mcData.protocol.states[state][direction];
   const proto=createProtocol(mcData.protocol.types,packets);
-  return new Parser(proto,"packet");
+  return new Parser(proto,"packet",customPackets);
 }
 
 module.exports = {